updated _putint.c

+++ _putint.c
@@ -28,7 +28,7 @@ return (res);
  */
 void print_number(int n)
 {
-unsigned int n1;
+unsigned long int n1;
 
 if (n < 0)
 {
